+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Toggle-disk can now be retried or aborted.
 - Added export-db API
 - Satellites report platform and OS variant and Controller exposes that on the REST API.
+- Added property "ExtCmd/WaitTimeout"
+- Added property "ExtCmd/TermTimeout"
+- Added property "ExtCmd/KillTimeout"
+- Added property "ExtCmd/IoStallTimeout"
+- Added property "ExtCmd/IoPollInterval"
 
 ### Changed